Added a working Crank Rocker with TwoBoneIK
FABRIK does not work great if you have poses which can be reached from multiple configurations (e.g. inwards and outwards bending). This can be circumvented with constrains but it's sometimes better to simply use TwoBoneIK if there are only 2 bones and you know if the movement can be described by only one bending.